separate bandwidth tests by number of instances
