--- deploy/workloads/bw-test/bw-test-client-0.riscv
@@ -1,1 +1,0 @@
-../../../sw/network-benchmarks/testbuild/bw-test-client-0.riscv
