--- deploy/workloads/bw-test/bw-test-server-8.riscv
@@ -1,1 +1,0 @@
-../../../sw/network-benchmarks/testbuild/bw-test-server-8.riscv
