--- deploy/workloads/bw-test/switchpatch.patch
@@ -1,37 +1,0 @@
-diff --git a/target-design/switch/baseport.h b/target-design/switch/baseport.h
-index 8733e52..cd42300 100644
---- a/target-design/switch/baseport.h
-+++ b/target-design/switch/baseport.h
-@@ -85,6 +85,8 @@ void BasePort::write_flits_to_output() {
-             switchpacket * thispacket = outputqueue.front();
-             outputqueue.pop();
-             printf("intended timestamp: %ld, actual timestamp: %ld, diff %ld\n", outputtimestamp, basetime + flitswritten, (int64_t)(basetime + flitswritten) - (int64_t)(outputtimestamp));
-+            // for root switch BW logging
-+            printf("packet timestamp: %ld, len: %ld\n", basetime + flitswritten, thispacket->amtwritten);
-             for (int i = 0; i < thispacket->amtwritten; i++) {
-                 write_last_flit(current_output_buf, flitswritten, i == (thispacket->amtwritten-1));
-                 write_valid_flit(current_output_buf, flitswritten);
-diff --git a/target-design/switch/switch.cc b/target-design/switch/switch.cc
-index 855a2ea..77ef270 100644
---- a/target-design/switch/switch.cc
-+++ b/target-design/switch/switch.cc
-@@ -11,7 +11,7 @@
- #include <omp.h>
- #include <cstdlib>
- 
--#define IGNORE_PRINTF
-+//#define IGNORE_PRINTF
- 
- #ifdef IGNORE_PRINTF
- #define printf(fmt, ...) (0)
-@@ -156,8 +156,8 @@ while (!pqueue.empty()) {
-     switchpacket * tsp = pqueue.top().switchpack;
-     pqueue.pop();
-     uint16_t send_to_port = get_port_from_flit(tsp->dat[0], 0 /* junk remove arg */);
--    printf("packet for port: %x\n", send_to_port);
--    printf("packet timestamp: %ld\n", tsp->timestamp);
-+    //printf("packet for port: %x\n", send_to_port);
-+    //printf("packet timestamp: %ld\n", tsp->timestamp);
-     if (send_to_port == BROADCAST_ADJUSTED) {
-         for (int i = 0; i < NUMPORTS; i++) {
-             if (i != tsp->sender ) {
